% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{desstat}
\alias{desstat}
\title{Descriptive statistical function using R}
\usage{
desstat(x, parametirc = TRUE, print = FALSE)
}
\arguments{
\item{x}{A set of data}

\item{parametirc}{Decide what parameters to output}

\item{print}{Print the results or not}
}
\value{
if parametirc=TRUE then output mean and variance,else output median absolute deviation and medianche
}
\description{
Output the mean and variance of a set of data, or the median absolute deviation and median
}
\examples{
\dontrun{
x<-rnorm(500)
y<- desstat(x,parametirc=TRUE,print=TRUE)
}
}
