% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Metropolis}
\alias{Metropolis}
\title{A generation of standard Laplace distribution using Rcpp}
\usage{
Metropolis(sigma, x0, N)
}
\arguments{
\item{sigma}{the parameter of proposal distribution}

\item{x0}{the initial values}

\item{N}{the number of samples}
}
\value{
a random sample of size \code{N}
}
\description{
A generation of standard Laplace distribution using Rcpp
}
\examples{
\dontrun{
LAPsamples <- Metropolis(.2,5,3000)
}
}
