% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{mu}
\alias{mu}
\title{Estimation of Confidence interval of population mean}
\usage{
mu(alpha, x, sigma = NA)
}
\arguments{
\item{alpha}{Significance level}

\item{x}{A set of data from a normal population}

\item{sigma}{Standard deviation of population distribution}
}
\value{
Two critical values and an interval
}
\description{
Estimate the confidence interval of mean of the normal population(The population variance is known)
}
\examples{
\dontrun{
x<-rnorm(10,mean=0,sd=1)
m<-mu(alpha=0.05,x,sigma=1)
m
}
}
