% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{mux_muy}
\alias{mux_muy}
\title{Estimation of confidence interval of the difference between the mean  of two normal populations}
\usage{
mux_muy(alpha, x, y, sigmax = NA, sigmay = NA)
}
\arguments{
\item{alpha}{Significance level}

\item{x}{A set of data from a normal population}

\item{y}{A set of data from a normal population}

\item{sigmax}{Standard deviation of population distribution x}

\item{sigmay}{Standard deviation of population distribution y}
}
\value{
a confidence interval
}
\description{
Estimate the confidence interval of the difference between the mean  of two normal populations(The population variance is known)
}
\examples{
\dontrun{
x<-rnorm(10,mean=0,sd=1)
y<-rnorm(10,mean=0,sd=1)
m<-mux_muy(alpha=0.05,x,y,sigmax=1,sigmay=1)
m
}
}
