\name{StatComp20012-package}
\alias{StatComp20012-package}
\alias{StatComp20012}
\docType{package}
\title{
  some functions for confidence interval estimation
}
\description{
  Several R functions are included in this package for confidence interval estimation.
}
\details{
  StatComp20012 is a simple R package developed for the conventional interval estimation of data from normal population.(\code{desstat} and \code{mu} and \code{sigmA} and \code{mux_muy})
}
\author{
Ye Peng

Maintainer: Ye Peng<1821293593@qq.com>
}

\keyword{ package }

